Split `http.py` into `repository` and `http_api_helpers`
In 'repository' we keep only AAS/Submodel/CD Repository App,
in `http_api_helpers` we keep all classes/funcs which
will be used across discovery/repository/registry apps
