+++ server/app/main.py
@@ -6,7 +6,7 @@ from basyx.aas import model, adapter
 from basyx.aas.adapter import aasx
 
 from basyx.aas.backend.local_file import LocalFileObjectStore
-from server.app.http import WSGIApp
+from server.app.repository import WSGIApp
 
 storage_path = os.getenv("STORAGE_PATH", "/storage")
 storage_type = os.getenv("STORAGE_TYPE", "LOCAL_FILE_READ_ONLY")
