+++ server/app/repository.py
@@ -1,9 +1,3 @@
-# Copyright (c) 2024 the Eclipse BaSyx Authors
-#
-# This program and the accompanying materials are made available under the terms of the MIT License, available in
-# the LICENSE file of this project.
-#
-# SPDX-License-Identifier: MIT
 """
 This module implements the "Specification of the Asset Administration Shell Part 2 Application Programming Interfaces".
 However, several features and routes are currently not supported:
@@ -34,442 +28,22 @@ However, several features and routes are currently not supported:
    - `GET /submodels/{submodelIdentifier}/submodel-elements/{idShortPath}/operation-results/{handleId}/$value`
 """
 
-import abc
-import base64
-import binascii
-import datetime
-import enum
 import io
-import json
 import itertools
+from typing import Iterable, Type, Iterator, List, Dict, Union, Callable, Tuple, Optional
 
-from lxml import etree
 import werkzeug.exceptions
 import werkzeug.routing
-import werkzeug.urls
 import werkzeug.utils
-from werkzeug.exceptions import BadRequest, Conflict, NotFound, UnprocessableEntity
-from werkzeug.routing import MapAdapter, Rule, Submount
-from werkzeug.wrappers import Request, Response
+from werkzeug import Response, Request
 from werkzeug.datastructures import FileStorage
+from werkzeug.exceptions import NotFound, BadRequest, Conflict
+from werkzeug.routing import Submount, Rule, MapAdapter
 
-import server.app.server_model as server_model
 from basyx.aas import model
-from basyx.aas.adapter._generic import XML_NS_MAP
-
 from basyx.aas.adapter import aasx
-
-from basyx.aas.adapter.xml import xml_serialization, XMLConstructables
-
-from server.app.adapter.xmlization import ServerXMLConstructables, read_server_aas_xml_element
-from server.app.adapter.jsonization import ServerAASToJsonEncoder, ServerStrictAASFromJsonDecoder, ServerStrictStrippedAASFromJsonDecoder
-
-from typing import Callable, Dict, Iterable, Iterator, List, Optional, Type, TypeVar, Union, Tuple
-
-
-@enum.unique
-class MessageType(enum.Enum):
-    UNDEFINED = enum.auto()
-    INFO = enum.auto()
-    WARNING = enum.auto()
-    ERROR = enum.auto()
-    EXCEPTION = enum.auto()
-
-    def __str__(self):
-        return self.name.capitalize()
-
-
-class Message:
-    def __init__(self, code: str, text: str, message_type: MessageType = MessageType.UNDEFINED,
-                 timestamp: Optional[datetime.datetime] = None):
-        self.code: str = code
-        self.text: str = text
-        self.message_type: MessageType = message_type
-        self.timestamp: datetime.datetime = timestamp if timestamp is not None \
-            else datetime.datetime.now(datetime.timezone.utc)
-
-
-class Result:
-    def __init__(self, success: bool, messages: Optional[List[Message]] = None):
-        if messages is None:
-            messages = []
-        self.success: bool = success
-        self.messages: List[Message] = messages
-
-
-class ResultToJsonEncoder(ServerAASToJsonEncoder):
-    @classmethod
-    def _result_to_json(cls, result: Result) -> Dict[str, object]:
-        return {
-            "success": result.success,
-            "messages": result.messages
-        }
-
-    @classmethod
-    def _message_to_json(cls, message: Message) -> Dict[str, object]:
-        return {
-            "messageType": message.message_type,
-            "text": message.text,
-            "code": message.code,
-            "timestamp": message.timestamp.isoformat()
-        }
-
-    def default(self, obj: object) -> object:
-        if isinstance(obj, Result):
-            return self._result_to_json(obj)
-        if isinstance(obj, Message):
-            return self._message_to_json(obj)
-        if isinstance(obj, MessageType):
-            return str(obj)
-        return super().default(obj)
-
-
-class StrippedResultToJsonEncoder(ResultToJsonEncoder):
-    stripped = True
-
-
-ResponseData = Union[Result, object, List[object]]
-
-
-class APIResponse(abc.ABC, Response):
-    @abc.abstractmethod
-    def __init__(self, obj: Optional[ResponseData] = None, cursor: Optional[int] = None,
-                 stripped: bool = False, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if obj is None:
-            self.status_code = 204
-        else:
-            self.data = self.serialize(obj, cursor, stripped)
-
-    @abc.abstractmethod
-    def serialize(self, obj: ResponseData, cursor: Optional[int], stripped: bool) -> str:
-        pass
-
-
-class JsonResponse(APIResponse):
-    def __init__(self, *args, content_type="application/json", **kwargs):
-        super().__init__(*args, **kwargs, content_type=content_type)
-
-    def serialize(self, obj: ResponseData, cursor: Optional[int], stripped: bool) -> str:
-        if cursor is None:
-            data = obj
-        else:
-            data = {
-                "paging_metadata": {"cursor": str(cursor)},
-                "result": obj
-            }
-        return json.dumps(
-            data,
-            cls=StrippedResultToJsonEncoder if stripped else ResultToJsonEncoder,
-            separators=(",", ":")
-        )
-
-
-class XmlResponse(APIResponse):
-    def __init__(self, *args, content_type="application/xml", **kwargs):
-        super().__init__(*args, **kwargs, content_type=content_type)
-
-    def serialize(self, obj: ResponseData, cursor: Optional[int], stripped: bool) -> str:
-        root_elem = etree.Element("response", nsmap=XML_NS_MAP)
-        if cursor is not None:
-            root_elem.set("cursor", str(cursor))
-        if isinstance(obj, Result):
-            result_elem = result_to_xml(obj, **XML_NS_MAP)
-            for child in result_elem:
-                root_elem.append(child)
-        elif isinstance(obj, list):
-            for item in obj:
-                item_elem = xml_serialization.object_to_xml_element(item)
-                root_elem.append(item_elem)
-        else:
-            obj_elem = xml_serialization.object_to_xml_element(obj)
-            for child in obj_elem:
-                root_elem.append(child)
-        etree.cleanup_namespaces(root_elem)
-        xml_str = etree.tostring(root_elem, xml_declaration=True, encoding="utf-8")
-        return xml_str  # type: ignore[return-value]
-
-
-class XmlResponseAlt(XmlResponse):
-    def __init__(self, *args, content_type="text/xml", **kwargs):
-        super().__init__(*args, **kwargs, content_type=content_type)
-
-
-def result_to_xml(result: Result, **kwargs) -> etree._Element:
-    result_elem = etree.Element("result", **kwargs)
-    success_elem = etree.Element("success")
-    success_elem.text = xml_serialization.boolean_to_xml(result.success)
-    messages_elem = etree.Element("messages")
-    for message in result.messages:
-        messages_elem.append(message_to_xml(message))
-
-    result_elem.append(success_elem)
-    result_elem.append(messages_elem)
-    return result_elem
-
-
-def message_to_xml(message: Message) -> etree._Element:
-    message_elem = etree.Element("message")
-    message_type_elem = etree.Element("messageType")
-    message_type_elem.text = str(message.message_type)
-    text_elem = etree.Element("text")
-    text_elem.text = message.text
-    code_elem = etree.Element("code")
-    code_elem.text = message.code
-    timestamp_elem = etree.Element("timestamp")
-    timestamp_elem.text = message.timestamp.isoformat()
-
-    message_elem.append(message_type_elem)
-    message_elem.append(text_elem)
-    message_elem.append(code_elem)
-    message_elem.append(timestamp_elem)
-    return message_elem
-
-
-def get_response_type(request: Request) -> Type[APIResponse]:
-    response_types: Dict[str, Type[APIResponse]] = {
-        "application/json": JsonResponse,
-        "application/xml": XmlResponse,
-        "text/xml": XmlResponseAlt
-    }
-    if len(request.accept_mimetypes) == 0 or request.accept_mimetypes.best in (None, "*/*"):
-        return JsonResponse
-    mime_type = request.accept_mimetypes.best_match(response_types)
-    if mime_type is None:
-        raise werkzeug.exceptions.NotAcceptable("This server supports the following content types: "
-                                                + ", ".join(response_types.keys()))
-    return response_types[mime_type]
-
-
-def http_exception_to_response(exception: werkzeug.exceptions.HTTPException, response_type: Type[APIResponse]) \
-        -> APIResponse:
-    headers = exception.get_headers()
-    location = exception.get_response().location
-    if location is not None:
-        headers.append(("Location", location))
-    if exception.code and exception.code >= 400:
-        message = Message(type(exception).__name__, exception.description if exception.description is not None else "",
-                          MessageType.ERROR)
-        result = Result(False, [message])
-    else:
-        result = Result(False)
-    return response_type(result, status=exception.code, headers=headers)
-
-
-def is_stripped_request(request: Request) -> bool:
-    return request.args.get("level") == "core"
-
-
-T = TypeVar("T")
-
-BASE64URL_ENCODING = "utf-8"
-
-
-def base64url_decode(data: str) -> str:
-    try:
-        # If the requester omits the base64 padding, an exception will be raised.
-        # However, Python doesn't complain about too much padding,
-        # thus we simply always append two padding characters (==).
-        # See also: https://stackoverflow.com/a/49459036/4780052
-        decoded = base64.urlsafe_b64decode(data + "==").decode(BASE64URL_ENCODING)
-    except binascii.Error:
-        raise BadRequest(f"Encoded data {data} is invalid base64url!")
-    except UnicodeDecodeError:
-        raise BadRequest(f"Encoded base64url value is not a valid {BASE64URL_ENCODING} string!")
-    return decoded
-
-
-def base64url_encode(data: str) -> str:
-    encoded = base64.urlsafe_b64encode(data.encode(BASE64URL_ENCODING)).decode("ascii")
-    return encoded
-
-
-class HTTPApiDecoder:
-    # these are the types we can construct (well, only the ones we need)
-    type_constructables_map = {
-        model.AssetAdministrationShell: XMLConstructables.ASSET_ADMINISTRATION_SHELL,
-        model.AssetInformation: XMLConstructables.ASSET_INFORMATION,
-        model.ModelReference: XMLConstructables.MODEL_REFERENCE,
-        model.SpecificAssetId: XMLConstructables.SPECIFIC_ASSET_ID,
-        model.Qualifier: XMLConstructables.QUALIFIER,
-        model.Submodel: XMLConstructables.SUBMODEL,
-        model.SubmodelElement: XMLConstructables.SUBMODEL_ELEMENT,
-        model.Reference: XMLConstructables.REFERENCE,
-        server_model.AssetAdministrationShellDescriptor: ServerXMLConstructables.ASSET_ADMINISTRATION_SHELL_DESCRIPTOR,
-        server_model.SubmodelDescriptor: ServerXMLConstructables.SUBMODEL_DESCRIPTOR,
-        server_model.AssetLink: ServerXMLConstructables.ASSET_LINK,
-    }
-
-    @classmethod
-    def check_type_supportance(cls, type_: type):
-        if type_ not in cls.type_constructables_map:
-            raise TypeError(f"Parsing {type_} is not supported!")
-
-    @classmethod
-    def assert_type(cls, obj: object, type_: Type[T]) -> T:
-        if not isinstance(obj, type_):
-            raise UnprocessableEntity(f"Object {obj!r} is not of type {type_.__name__}!")
-        return obj
-
-    @classmethod
-    def json_list(cls, data: Union[str, bytes], expect_type: Type[T], stripped: bool, expect_single: bool) -> List[T]:
-        cls.check_type_supportance(expect_type)
-        decoder: Type[ServerStrictAASFromJsonDecoder] = ServerStrictStrippedAASFromJsonDecoder if stripped \
-            else ServerStrictAASFromJsonDecoder
-        try:
-            parsed = json.loads(data, cls=decoder)
-            if not isinstance(parsed, list):
-                if not expect_single:
-                    raise UnprocessableEntity(f"Expected List[{expect_type.__name__}], got {parsed!r}!")
-                parsed = [parsed]
-            elif expect_single:
-                raise UnprocessableEntity(f"Expected a single object of type {expect_type.__name__}, got {parsed!r}!")
-            # TODO: the following is ugly, but necessary because references aren't self-identified objects
-            #  in the json schema
-            # TODO: json deserialization will always create an ModelReference[Submodel], xml deserialization determines
-            #  that automatically
-            constructor: Optional[Callable[..., T]] = None
-            args = []
-            if expect_type is model.ModelReference:
-                constructor = decoder._construct_model_reference  # type: ignore[assignment]
-                args.append(model.Submodel)
-            elif expect_type is model.AssetInformation:
-                constructor = decoder._construct_asset_information  # type: ignore[assignment]
-            elif expect_type is model.SpecificAssetId:
-                constructor = decoder._construct_specific_asset_id  # type: ignore[assignment]
-            elif expect_type is model.Reference:
-                constructor = decoder._construct_reference  # type: ignore[assignment]
-            elif expect_type is model.Qualifier:
-                constructor = decoder._construct_qualifier  # type: ignore[assignment]
-            elif expect_type is model.AssetAdministrationShellDescriptor:
-                constructor = decoder._construct_asset_administration_shell_descriptor
-            elif expect_type is model.SubmodelDescriptor:
-                constructor = decoder._construct_submodel_descriptor
-            elif expect_type is server_model.AssetLink:
-                constructor = decoder._construct_asset_link
-
-            if constructor is not None:
-                # construct elements that aren't self-identified
-                return [constructor(obj, *args) for obj in parsed]
-
-        except (KeyError, ValueError, TypeError, json.JSONDecodeError, model.AASConstraintViolation) as e:
-            raise UnprocessableEntity(str(e)) from e
-
-        return [cls.assert_type(obj, expect_type) for obj in parsed]
-
-    @classmethod
-    def base64urljson_list(cls, data: str, expect_type: Type[T], stripped: bool, expect_single: bool) -> List[T]:
-        data = base64url_decode(data)
-        return cls.json_list(data, expect_type, stripped, expect_single)
-
-    @classmethod
-    def json(cls, data: Union[str, bytes], expect_type: Type[T], stripped: bool) -> T:
-        return cls.json_list(data, expect_type, stripped, True)[0]
-
-    @classmethod
-    def base64urljson(cls, data: str, expect_type: Type[T], stripped: bool) -> T:
-        data = base64url_decode(data)
-        return cls.json_list(data, expect_type, stripped, True)[0]
-
-    @classmethod
-    def xml(cls, data: bytes, expect_type: Type[T], stripped: bool) -> T:
-        cls.check_type_supportance(expect_type)
-        try:
-            xml_data = io.BytesIO(data)
-            rv = read_server_aas_xml_element(xml_data, cls.type_constructables_map[expect_type],
-                                             stripped=stripped, failsafe=False)
-        except (KeyError, ValueError) as e:
-            # xml deserialization creates an error chain. since we only return one error, return the root cause
-            f: BaseException = e
-            while f.__cause__ is not None:
-                f = f.__cause__
-            raise UnprocessableEntity(str(f)) from e
-        except (etree.XMLSyntaxError, model.AASConstraintViolation) as e:
-            raise UnprocessableEntity(str(e)) from e
-        return cls.assert_type(rv, expect_type)
-
-    @classmethod
-    def request_body(cls, request: Request, expect_type: Type[T], stripped: bool) -> T:
-        """
-        TODO: werkzeug documentation recommends checking the content length before retrieving the body to prevent
-              running out of memory. but it doesn't state how to check the content length
-              also: what would be a reasonable maximum content length? the request body isn't limited by the xml/json
-              schema
-            In the meeting (25.11.2020) we discussed, this may refer to a reverse proxy in front of this WSGI app,
-            which should limit the maximum content length.
-        """
-        valid_content_types = ("application/json", "application/xml", "text/xml")
-
-        if request.mimetype not in valid_content_types:
-            raise werkzeug.exceptions.UnsupportedMediaType(
-                f"Invalid content-type: {request.mimetype}! Supported types: "
-                + ", ".join(valid_content_types))
-
-        if request.mimetype == "application/json":
-            return cls.json(request.get_data(), expect_type, stripped)
-        return cls.xml(request.get_data(), expect_type, stripped)
-    @classmethod
-    def request_body_list(cls, request: Request, expect_type: Type[T], stripped: bool) -> T:
-        """
-        Deserializes the request body to an instance (or list of instances)
-        of the expected type.
-        """
-        valid_content_types = ("application/json", "application/xml", "text/xml")
-
-        if request.mimetype not in valid_content_types:
-            raise werkzeug.exceptions.UnsupportedMediaType(
-                f"Invalid content-type: {request.mimetype}! Supported types: " + ", ".join(valid_content_types)
-            )
-
-        if request.mimetype == "application/json":
-            raw_data = request.get_data()
-            try:
-                parsed = json.loads(raw_data)
-            except Exception as e:
-                raise werkzeug.exceptions.BadRequest(f"Invalid JSON: {e}")
-            # Prüfe, ob parsed ein Array ist:
-            if isinstance(parsed, list):
-                # Für jedes Element wird die Konvertierung angewandt.
-                return [cls._convert_single_json_item(item, expect_type, stripped) for item in parsed]  # type: ignore
-            else:
-                return cls._convert_single_json_item(parsed, expect_type, stripped)
-        else:
-            return cls.xml(request.get_data(), expect_type, stripped)
-
-    @classmethod
-    def _convert_single_json_item(cls, data: any, expect_type: Type[T], stripped: bool) -> T:
-        """
-        Konvertiert ein einzelnes JSON-Objekt (als Python-Dict) in ein Objekt vom Typ expect_type.
-        Hierbei wird das Dictionary zuerst wieder in einen JSON-String serialisiert und als Bytes übergeben.
-        """
-        json_bytes = json.dumps(data).encode("utf-8")
-        return cls.json(json_bytes, expect_type, stripped)
-
-class Base64URLConverter(werkzeug.routing.UnicodeConverter):
-
-    def to_url(self, value: model.Identifier) -> str:
-        return super().to_url(base64url_encode(value))
-
-    def to_python(self, value: str) -> model.Identifier:
-        value = super().to_python(value)
-        decoded = base64url_decode(super().to_python(value))
-        return decoded
-
-
-class IdShortPathConverter(werkzeug.routing.UnicodeConverter):
-    id_short_sep = "."
-
-    def to_url(self, value: List[str]) -> str:
-        return super().to_url(self.id_short_sep.join(value))
-
-    def to_python(self, value: str) -> List[str]:
-        id_shorts = super().to_python(value).split(self.id_short_sep)
-        for id_short in id_shorts:
-            try:
-                model.Referable.validate_id_short(id_short)
-            except (ValueError, model.AASConstraintViolation):
-                raise BadRequest(f"{id_short} is not a valid id_short!")
-        return id_shorts
+from server.app.http_api_helpers import Base64URLConverter, IdShortPathConverter, T, HTTPApiDecoder, get_response_type, \
+    http_exception_to_response, APIResponse, is_stripped_request
 
 
 class WSGIApp:
@@ -492,17 +66,13 @@ class WSGIApp:
                         Rule("/$reference", methods=["GET"], endpoint=self.get_aas_reference),
                         Rule("/asset-information", methods=["GET"], endpoint=self.get_aas_asset_information),
                         Rule("/asset-information", methods=["PUT"], endpoint=self.put_aas_asset_information),
-                        Rule("/asset-information/thumbnail", methods=["GET", "PUT", "DELETE"],
-                             endpoint=self.not_implemented),
+                        Rule("/asset-information/thumbnail", methods=["GET", "PUT", "DELETE"], endpoint=self.not_implemented),
                         Rule("/submodel-refs", methods=["GET"], endpoint=self.get_aas_submodel_refs),
                         Rule("/submodel-refs", methods=["POST"], endpoint=self.post_aas_submodel_refs),
-                        Rule("/submodel-refs/<base64url:submodel_id>", methods=["DELETE"],
-                             endpoint=self.delete_aas_submodel_refs_specific),
+                        Rule("/submodel-refs/<base64url:submodel_id>", methods=["DELETE"], endpoint=self.delete_aas_submodel_refs_specific),
                         Submount("/submodels", [
-                            Rule("/<base64url:submodel_id>", methods=["PUT"],
-                                 endpoint=self.put_aas_submodel_refs_submodel),
-                            Rule("/<base64url:submodel_id>", methods=["DELETE"],
-                                 endpoint=self.delete_aas_submodel_refs_submodel),
+                            Rule("/<base64url:submodel_id>", methods=["PUT"], endpoint=self.put_aas_submodel_refs_submodel),
+                            Rule("/<base64url:submodel_id>", methods=["DELETE"], endpoint=self.delete_aas_submodel_refs_submodel),
                             Rule("/<base64url:submodel_id>", endpoint=self.aas_submodel_refs_redirect),
                             Rule("/<base64url:submodel_id>/<path:path>", endpoint=self.aas_submodel_refs_redirect)
                         ])
@@ -527,76 +97,51 @@ class WSGIApp:
                         Rule("/$reference", methods=["GET"], endpoint=self.get_submodels_reference),
                         Rule("/$path", methods=["GET"], endpoint=self.not_implemented),
                         Rule("/submodel-elements", methods=["GET"], endpoint=self.get_submodel_submodel_elements),
-                        Rule("/submodel-elements", methods=["POST"],
-                             endpoint=self.post_submodel_submodel_elements_id_short_path),
+                        Rule("/submodel-elements", methods=["POST"], endpoint=self.post_submodel_submodel_elements_id_short_path),
                         Submount("/submodel-elements", [
-                            Rule("/$metadata", methods=["GET"],
-                                 endpoint=self.get_submodel_submodel_elements_metadata),
-                            Rule("/$reference", methods=["GET"],
-                                 endpoint=self.get_submodel_submodel_elements_reference),
+                            Rule("/$metadata", methods=["GET"], endpoint=self.get_submodel_submodel_elements_metadata),
+                            Rule("/$reference", methods=["GET"], endpoint=self.get_submodel_submodel_elements_reference),
                             Rule("/$value", methods=["GET"], endpoint=self.not_implemented),
                             Rule("/$path", methods=["GET"], endpoint=self.not_implemented),
-                            Rule("/<id_short_path:id_shorts>", methods=["GET"],
-                                 endpoint=self.get_submodel_submodel_elements_id_short_path),
-                            Rule("/<id_short_path:id_shorts>", methods=["POST"],
-                                 endpoint=self.post_submodel_submodel_elements_id_short_path),
-                            Rule("/<id_short_path:id_shorts>", methods=["PUT"],
-                                 endpoint=self.put_submodel_submodel_elements_id_short_path),
-                            Rule("/<id_short_path:id_shorts>", methods=["DELETE"],
-                                 endpoint=self.delete_submodel_submodel_elements_id_short_path),
+                            Rule("/<id_short_path:id_shorts>", methods=["GET"], endpoint=self.get_submodel_submodel_elements_id_short_path),
+                            Rule("/<id_short_path:id_shorts>", methods=["POST"], endpoint=self.post_submodel_submodel_elements_id_short_path),
+                            Rule("/<id_short_path:id_shorts>", methods=["PUT"], endpoint=self.put_submodel_submodel_elements_id_short_path),
+                            Rule("/<id_short_path:id_shorts>", methods=["DELETE"], endpoint=self.delete_submodel_submodel_elements_id_short_path),
                             Rule("/<id_short_path:id_shorts>", methods=["PATCH"], endpoint=self.not_implemented),
                             Submount("/<id_short_path:id_shorts>", [
-                                Rule("/$metadata", methods=["GET"],
-                                     endpoint=self.get_submodel_submodel_elements_id_short_path_metadata),
+                                Rule("/$metadata", methods=["GET"], endpoint=self.get_submodel_submodel_elements_id_short_path_metadata),
                                 Rule("/$metadata", methods=["PATCH"], endpoint=self.not_implemented),
-                                Rule("/$reference", methods=["GET"],
-                                     endpoint=self.get_submodel_submodel_elements_id_short_path_reference),
+                                Rule("/$reference", methods=["GET"], endpoint=self.get_submodel_submodel_elements_id_short_path_reference),
                                 Rule("/$value", methods=["GET"], endpoint=self.not_implemented),
                                 Rule("/$value", methods=["PATCH"], endpoint=self.not_implemented),
                                 Rule("/$path", methods=["GET"], endpoint=self.not_implemented),
-                                Rule("/attachment", methods=["GET"],
-                                     endpoint=self.get_submodel_submodel_element_attachment),
-                                Rule("/attachment", methods=["PUT"],
-                                     endpoint=self.put_submodel_submodel_element_attachment),
-                                Rule("/attachment", methods=["DELETE"],
-                                     endpoint=self.delete_submodel_submodel_element_attachment),
+                                Rule("/attachment", methods=["GET"], endpoint=self.get_submodel_submodel_element_attachment),
+                                Rule("/attachment", methods=["PUT"], endpoint=self.put_submodel_submodel_element_attachment),
+                                Rule("/attachment", methods=["DELETE"], endpoint=self.delete_submodel_submodel_element_attachment),
                                 Rule("/invoke", methods=["POST"], endpoint=self.not_implemented),
                                 Rule("/invoke/$value", methods=["POST"], endpoint=self.not_implemented),
                                 Rule("/invoke-async", methods=["POST"], endpoint=self.not_implemented),
                                 Rule("/invoke-async/$value", methods=["POST"], endpoint=self.not_implemented),
-                                Rule("/operation-status/<base64url:handleId>", methods=["GET"],
-                                     endpoint=self.not_implemented),
+                                Rule("/operation-status/<base64url:handleId>", methods=["GET"], endpoint=self.not_implemented),
                                 Submount("/operation-results", [
-                                    Rule("/<base64url:handleId>", methods=["GET"],
-                                         endpoint=self.not_implemented),
-                                    Rule("/<base64url:handleId>/$value", methods=["GET"],
-                                         endpoint=self.not_implemented)
+                                    Rule("/<base64url:handleId>", methods=["GET"], endpoint=self.not_implemented),
+                                    Rule("/<base64url:handleId>/$value", methods=["GET"], endpoint=self.not_implemented)
                                 ]),
-                                Rule("/qualifiers", methods=["GET"],
-                                     endpoint=self.get_submodel_submodel_element_qualifiers),
-                                Rule("/qualifiers", methods=["POST"],
-                                     endpoint=self.post_submodel_submodel_element_qualifiers),
+                                Rule("/qualifiers", methods=["GET"], endpoint=self.get_submodel_submodel_element_qualifiers),
+                                Rule("/qualifiers", methods=["POST"], endpoint=self.post_submodel_submodel_element_qualifiers),
                                 Submount("/qualifiers", [
-                                    Rule("/<base64url:qualifier_type>", methods=["GET"],
-                                         endpoint=self.get_submodel_submodel_element_qualifiers),
-                                    Rule("/<base64url:qualifier_type>", methods=["PUT"],
-                                         endpoint=self.put_submodel_submodel_element_qualifiers),
-                                    Rule("/<base64url:qualifier_type>", methods=["DELETE"],
-                                         endpoint=self.delete_submodel_submodel_element_qualifiers)
+                                    Rule("/<base64url:qualifier_type>", methods=["GET"], endpoint=self.get_submodel_submodel_element_qualifiers),
+                                    Rule("/<base64url:qualifier_type>", methods=["PUT"], endpoint=self.put_submodel_submodel_element_qualifiers),
+                                    Rule("/<base64url:qualifier_type>", methods=["DELETE"], endpoint=self.delete_submodel_submodel_element_qualifiers)
                                 ])
                             ])
                         ]),
-                        Rule("/qualifiers", methods=["GET"],
-                             endpoint=self.get_submodel_submodel_element_qualifiers),
-                        Rule("/qualifiers", methods=["POST"],
-                             endpoint=self.post_submodel_submodel_element_qualifiers),
+                        Rule("/qualifiers", methods=["GET"], endpoint=self.get_submodel_submodel_element_qualifiers),
+                        Rule("/qualifiers", methods=["POST"], endpoint=self.post_submodel_submodel_element_qualifiers),
                         Submount("/qualifiers", [
-                            Rule("/<base64url:qualifier_type>", methods=["GET"],
-                                 endpoint=self.get_submodel_submodel_element_qualifiers),
-                            Rule("/<base64url:qualifier_type>", methods=["PUT"],
-                                 endpoint=self.put_submodel_submodel_element_qualifiers),
-                            Rule("/<base64url:qualifier_type>", methods=["DELETE"],
-                                 endpoint=self.delete_submodel_submodel_element_qualifiers)
+                            Rule("/<base64url:qualifier_type>", methods=["GET"], endpoint=self.get_submodel_submodel_element_qualifiers),
+                            Rule("/<base64url:qualifier_type>", methods=["PUT"], endpoint=self.put_submodel_submodel_element_qualifiers),
+                            Rule("/<base64url:qualifier_type>", methods=["DELETE"], endpoint=self.delete_submodel_submodel_element_qualifiers)
                         ])
                     ])
                 ]),
@@ -1209,7 +754,6 @@ class WSGIApp:
                                    **_kwargs) -> Response:
         self.object_store.remove(self._get_concept_description(url_args))
         return response_t()
-
 
 if __name__ == "__main__":
     from werkzeug.serving import run_simple
